moved functions to small modules
