+++ src/offset.ts
@@ -1,0 +1,11 @@
+/**
+ * @param {Element} element
+ * @returns {{left: *, top: *}}
+ */
+export default function (element) {
+  var rect = element.getClientRects()[0]
+  return {
+    left: rect.left + window.scrollX,
+    top: rect.top + window.scrollY
+  }
+}
